+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ## Changed
 - Fix undefined index warning parsing origin URL in ManifestFactory.
+- Fix third level manifests not inheriting first level tokens.
 
 ## 1.1.0
 ### Added